Do not wp_die for plugin admin page

+++ wa-rsfp.php
@@ -31,11 +31,11 @@ if ( ! defined( 'WPINC' ) ) {
 	die;
 }
 
-if ( is_admin() && !function_exists('rwmb_meta') ) {
+if ( is_admin() && !in_array( get_current_screen()->id, array( 'plugins.php' )) && !function_exists('rwmb_meta') ) {
 	wp_die('Error : please install Meta Box plugin.');
 }
 
-if ( is_admin() && !function_exists('mb_term_meta_load') ) {
+if ( is_admin() && !in_array( get_current_screen()->id, array( 'plugins.php' )) && !function_exists('mb_term_meta_load') ) {
 	wp_die('Error : please install Meta Box Term meta plugin.');
 }
